Disable Line numbers for code blocks, because of buggy rendering
With activated line numbers, there is a spacing between the line
number and the code, which is random for every new code block.

As I don't know the reason for that, and there is no fix in sight,
I'm just disabling this for now (which is unfortunate, because
being able to reference to a certain line number is nice for some
tutorials but anyways..).